lib: Add mountutil module
I copy/pasted the mountpoint check into bootc, but it'd
be better to centralize.
