+++ lib/src/commit.rs
@@ -3,13 +3,13 @@
 //! <https://github.com/ostreedev/ostree-rs-ext/issues/159>
 
 use crate::container_utils::require_ostree_container;
+use crate::mountutil::is_mountpoint;
 use anyhow::Context;
 use anyhow::Result;
 use camino::Utf8Path;
 use cap_std::fs::Dir;
 use cap_std_ext::cap_std;
 use cap_std_ext::dirext::CapStdExtDirExt;
-use io_lifetimes::AsFd;
 use rustix::fs::MetadataExt;
 use std::borrow::Cow;
 use std::convert::TryInto;
@@ -102,28 +102,6 @@ fn remove_all_on_mount_recurse(root: &Dir, rootdev: u64, path: &Path) -> Result<
     Ok(skipped)
 }
 
-/// Try to (heuristically) determine if the provided path is a mount root.
-fn is_mountpoint(root: &Dir, path: &Path) -> Result<Option<bool>> {
-    // https://github.com/systemd/systemd/blob/8fbf0a214e2fe474655b17a4b663122943b55db0/src/basic/mountpoint-util.c#L176
-    use rustix::fs::{AtFlags, StatxFlags};
-
-    // SAFETY(unwrap): We can infallibly convert an i32 into a u64.
-    let mountroot_flag: u64 = libc::STATX_ATTR_MOUNT_ROOT.try_into().unwrap();
-    match rustix::fs::statx(
-        root.as_fd(),
-        path,
-        AtFlags::NO_AUTOMOUNT | AtFlags::SYMLINK_NOFOLLOW,
-        StatxFlags::empty(),
-    ) {
-        Ok(r) => {
-            let present = (r.stx_attributes_mask & mountroot_flag) > 0;
-            Ok(present.then(|| r.stx_attributes & mountroot_flag > 0))
-        }
-        Err(e) if e == rustix::io::Errno::NOSYS => Ok(None),
-        Err(e) => Err(e.into()),
-    }
-}
-
 fn clean_subdir(root: &Dir, rootdev: u64) -> Result<()> {
     for entry in root.entries()? {
         let entry = entry?;
@@ -206,20 +184,6 @@ pub(crate) async fn container_commit() -> Result<()> {
 mod tests {
     use super::*;
     use cap_std_ext::cap_tempfile;
-
-    #[test]
-    fn test_is_mountpoint() -> Result<()> {
-        let root = cap_std::fs::Dir::open_ambient_dir("/", cap_std::ambient_authority())?;
-        let supported = is_mountpoint(&root, Path::new("/")).unwrap();
-        match supported {
-            Some(r) => assert!(r),
-            // If the host doesn't support statx, ignore this for now
-            None => return Ok(()),
-        }
-        let tmpdir = cap_tempfile::TempDir::new(cap_std::ambient_authority())?;
-        assert!(!is_mountpoint(&tmpdir, Path::new(".")).unwrap().unwrap());
-        Ok(())
-    }
 
     #[test]
     fn commit() -> Result<()> {
